# Reuse char[] when constructing trace context header for injection.