--- a/api/src/main/java/io/opentelemetry/internal/TemporaryBuffers.java
+++ b/api/src/main/java/io/opentelemetry/internal/TemporaryBuffers.java
@@ -15,8 +15,7 @@ public final class TemporaryBuffers {
    * A {@link ThreadLocal} {@code char[]} of size {@code len}. Take care when using a large value of
    * {@code len} as this buffer will remain for the lifetime of the thread. The returned buffer will
    * not be zeroed and may be larger than the requested size, you must make sure to fill the entire
-   * content to the desired value and set the length explicitly when converting to a
-   * {@link String}.
+   * content to the desired value and set the length explicitly when converting to a {@link String}.
    */
   public static char[] chars(int len) {
     char[] buffer = CHAR_ARRAY.get();