For #1430 I will probably add a similar reusable `StringBuilder` for reducing overhead of recording exceptions. First I figured we should apply it to our existing place to understand the implications.

When constructing strings, it is almost always worth using a reused buffer which avoids zeroing the array and extra garbage generation, and this works well since these buffers do not get exposed outside of the method.

Main concern is putting this in `internal`, we should reduce the number of these `ThreadLocal` that we have in an app since there's no point at all in having multiple, but with `internal` it's not supposed to be used outside, e.g., in the SDK or in propagation extensions. But in `common` it seems to expose it to the public. Does it make sense to allow this sort of class outside the API module? I can add the error prone `RestrictedApi` annotation to formalize that if it does. It's too bad the JDK will never add this sort of class since it's somewhat unsafe, and JDK never exposes unsafe APIs to users. But memory usage of apps would go down if there was one thread local to rule them all...

About 15% faster with less than half of the allocation.
After
```
Benchmark                                                                       Mode  Cnt    Score     Error   Units
HttpTraceContextInjectBenchmark.measureInject                                   avgt   15   70.617 ▒   7.677   ns/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.alloc.rate                    avgt   15  868.646 ▒  82.041  MB/sec
HttpTraceContextInjectBenchmark.measureInject:╖gc.alloc.rate.norm               avgt   15   96.009 ▒   0.001    B/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Eden_Space           avgt   15  877.322 ▒ 109.523  MB/sec
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Eden_Space.norm      avgt   15   97.044 ▒   8.893    B/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Survivor_Space       avgt   15    0.004 ▒   0.003  MB/sec
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Survivor_Space.norm  avgt   15   ? 10?│              B/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.count                         avgt   15   76.000            counts
HttpTraceContextInjectBenchmark.measureInject:╖gc.time                          avgt   15  123.000                ms
```

Before
```
Benchmark                                                                       Mode  Cnt     Score     Error   Units
HttpTraceContextInjectBenchmark.measureInject                                   avgt   15    84.904 ▒  12.105   ns/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.alloc.rate                    avgt   15  1697.155 ▒ 213.918  MB/sec
HttpTraceContextInjectBenchmark.measureInject:╖gc.alloc.rate.norm               avgt   15   224.017 ▒   0.001    B/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Eden_Space           avgt   15  1692.461 ▒ 228.890  MB/sec
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Eden_Space.norm      avgt   15   223.377 ▒  10.128    B/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Survivor_Space       avgt   15     0.004 ▒   0.003  MB/sec
HttpTraceContextInjectBenchmark.measureInject:╖gc.churn.G1_Survivor_Space.norm  avgt   15     0.001 ▒   0.001    B/op
HttpTraceContextInjectBenchmark.measureInject:╖gc.count                         avgt   15   114.000            counts
HttpTraceContextInjectBenchmark.measureInject:╖gc.time                          avgt   15   205.000                ms
```