@anuraaga 2020-07-20
Yeah they're in the JMH folders. This one had an existing one so it's not in this PR

https://github.com/open-telemetry/opentelemetry-java/blob/master/api/src/jmh/java/io/opentelemetry/trace/propagation/HttpTraceContextInjectBenchmark.java